web_ui_fixed/requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
pymysql==1.1.1
transformers==4.41.2
torch==2.3.1
Pillow==10.3.0
openpyxl==3.1.5
pandas==2.2.2
python-dotenv==1.0.1
